+++ src/extension.ts
@@ -249,6 +249,7 @@ class ConfigData {
     workspacefolder: string;
     config: vscode.WorkspaceConfiguration;
     compileCommandsData: CompileCommandsData;
+    compileCommandsJsonPath: string = "";
 
     constructor(workspacefolder: string) {
         this.workspacefolder = workspacefolder;
@@ -295,9 +296,30 @@ class ConfigData {
     }
 
     compileCommandsJson(): string {
-        let compileCommandsJsonDefault = "${workspaceFolder}/compile_commands.json";
-        let compileCommandsJson = this.config.get("compile_commands", compileCommandsJsonDefault);
-        return this.replaceWorkspaceVars(compileCommandsJson);
+        let compileCommandsJsonDefault = "auto";
+        let compileCommandsJsonPath = this.config.get("compile_commands", compileCommandsJsonDefault);
+        const tests: readonly string[] = compileCommandsJsonPath === compileCommandsJsonDefault ?
+            [
+                "${workspaceFolder}/compile_commands.json",
+                "${workspaceFolder}/build/compile_commands.json",
+                "${fileWorkspaceFolder}/compile_commands.json",
+                "${fileWorkspaceFolder}/build/compile_commands.json",
+            ]
+            : [compileCommandsJsonPath];
+        for (let test of tests) {
+            try {
+                test = this.replaceWorkspaceVars(test);
+                fs.statSync(test);
+                compileCommandsJsonPath = test;
+                break;
+            }
+            catch (err) {
+                // Ignore, caught later.
+            }
+        }
+        log(DEBUG, "Using compileCommandsJson = '" + compileCommandsJsonPath + "'.");
+        this.compileCommandsJsonPath = compileCommandsJsonPath;
+        return this.compileCommandsJsonPath;
     }
 
     updateConfig() {
@@ -305,10 +327,11 @@ class ConfigData {
     }
 
     updateCompileCommands() {
+        let compileCommandsJsonLast = this.compileCommandsJsonPath;
         let compileCommandsJson = this.compileCommandsJson();
         try {
-            let stats = fs.statSync(compileCommandsJson);
-            if (stats.mtimeMs !== this.compileCommandsData.mtimeMs) {
+            let stats = fs.statSync(this.compileCommandsJsonPath);
+            if (stats.mtimeMs !== this.compileCommandsData.mtimeMs || compileCommandsJsonLast !== this.compileCommandsJsonPath) {
                 this.compileCommandsData = this.parseCompileCommands();
             }
         }